+++ src/main/java/net/tylermurphy/hideAndSeek/command/Join.java
@@ -5,7 +5,7 @@
  *
  * Kenshins Hide and Seek free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * he Free Software Foundation, either version 3 of the License.
+ * he Free Software Foundation version 3.
  *
  * Kenshins Hide and Seek is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -23,15 +23,9 @@ import static net.tylermurphy.hideAndSeek.configuration.Config.*;
 
 import net.tylermurphy.hideAndSeek.game.Board;
 import net.tylermurphy.hideAndSeek.game.Game;
-import net.tylermurphy.hideAndSeek.util.Status;
 import org.bukkit.Bukkit;
-import org.bukkit.GameMode;
-import org.bukkit.Location;
-import org.bukkit.attribute.Attribute;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-
-import java.util.Objects;
 
 import static net.tylermurphy.hideAndSeek.configuration.Localization.*;
 
@@ -52,29 +46,7 @@ public class Join implements ICommand {
 			return;
 		}
 
-		join(player);
-	}
-
-	public static void join(Player player){
-		if(Game.status == Status.STANDBY) {
-			player.getInventory().clear();
-			Board.addHider(player);
-			if(announceMessagesToNonPlayers) Bukkit.broadcastMessage(messagePrefix + message("GAME_JOIN").addPlayer(player));
-			else Game.broadcastMessage(messagePrefix + message("GAME_JOIN").addPlayer(player));
-			player.teleport(new Location(Bukkit.getWorld(lobbyWorld), lobbyPosition.getX(),lobbyPosition.getY(),lobbyPosition.getZ()));
-			player.setGameMode(GameMode.ADVENTURE);
-			Board.createLobbyBoard(player);
-			Board.reloadLobbyBoards();
-		} else {
-			Board.addSpectator(player);
-			player.sendMessage(messagePrefix + message("GAME_JOIN_SPECTATOR"));
-			player.setGameMode(GameMode.SPECTATOR);
-			Board.createGameBoard(player);
-			player.teleport(new Location(Bukkit.getWorld("hideandseek_"+spawnWorld), spawnPosition.getX(),spawnPosition.getY(),spawnPosition.getZ()));
-		}
-
-		player.setFoodLevel(20);
-		player.setHealth(Objects.requireNonNull(player.getAttribute(Attribute.GENERIC_MAX_HEALTH)).getBaseValue());
+		Game.join(player);
 	}
 
 	public String getLabel() {
