+++ src/main/java/net/tylermurphy/hideAndSeek/command/Wins.java
@@ -5,7 +5,7 @@
  *
  * Kenshins Hide and Seek free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * he Free Software Foundation, either version 3 of the License.
+ * he Free Software Foundation version 3.
  *
  * Kenshins Hide and Seek is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
